reorganize file hierarchy
